parse errors split: known and unknown
