+++ src/utils/__tests__/parse.test.ts
@@ -1,7 +1,7 @@
-import parse from './parse';
+import parse from '../parse';
 
 // Constants
-import { ERROR_PARSING } from '../constants/errors';
+import { ERROR_NETWORK, ERROR_PARSING } from '../../constants/errors';
 
 describe('src/utils/parse.ts', () => {
     describe('parse', () => {
@@ -9,7 +9,13 @@ describe('src/utils/parse.ts', () => {
             const fn = data => data;
             expect(parse(fn)(1)).toBe(1);
         });
-        it('error case', () => {
+        it('error case (known)', () => {
+            const fn = data => {
+                throw ERROR_NETWORK;
+            };
+            expect(() => parse(fn)(1)).toThrow(ERROR_NETWORK);
+        });
+        it('error case (UNknown)', () => {
             const fn = data => {
                 throw new Error(data);
             };
