+++ src/utils/parse.ts
@@ -1,10 +1,14 @@
 // Constants
-import { ERROR_PARSING } from '../constants/errors';
+import * as ERROR from '../constants/errors';
 
 export default fn => data => {
     try {
         return fn(data);
     } catch (error) {
-        throw ERROR_PARSING;
+        const knownErrors = Object.values(ERROR);
+        if (knownErrors.includes(error)) {
+            throw error;
+        }
+        throw ERROR.ERROR_PARSING;
     }
 };
